Split up relaunch fc banner component
