+++ src/components/v2v3/V2V3Project/banners/RelaunchFundingCycleBanner/index.ts
@@ -1,0 +1,1 @@
+export { RelaunchFundingCycleBanner } from './RelaunchFundingCycleBanner'
